requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
pytorch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
pillow==9.5.0
pyttsx3==2.90  # 用于声音报警